integration_test/viss/requirements.txt
asyncio
grpcio
protobuf
pytest
pytest-ordering
pytest-bdd
allure-pytest-bdd
requests
websocket-client
paho-mqtt
tinydb
